separate schemas into their own files
