+++ lib/collection.js
@@ -16,7 +16,6 @@ const ChangeStream = require('./change_stream');
 const executeOperation = require('./utils').executeOperation;
 const applyWriteConcern = require('./utils').applyWriteConcern;
 const resolveReadPreference = require('./utils').resolveReadPreference;
-const ClientSession = require('mongodb-core').Sessions.ClientSession;
 const validate = require('./options_validator').validate;
 const applyDefaults = require('./options_validator').applyDefaults;
 
@@ -63,6 +62,51 @@ const stats = require('./operations/collection_ops').stats;
 const updateDocuments = require('./operations/collection_ops').updateDocuments;
 const updateMany = require('./operations/collection_ops').updateMany;
 const updateOne = require('./operations/collection_ops').updateOne;
+
+// Schemas
+const aggregateSchema = require('./schemas/collection_schemas').aggregateSchema;
+const bulkWriteSchema = require('./schemas/collection_schemas').bulkWriteSchema;
+const countDocumentsSchema = require('./schemas/collection_schemas').countDocumentsSchema;
+const createIndexSchema = require('./schemas/collection_schemas').createIndexSchema;
+const createIndexesSchema = require('./schemas/collection_schemas').createIndexesSchema;
+const deleteManySchema = require('./schemas/collection_schemas').deleteManySchema;
+const deleteOneSchema = require('./schemas/collection_schemas').deleteOneSchema;
+const distinctSchema = require('./schemas/collection_schemas').distinctSchema;
+const dropSchema = require('./schemas/collection_schemas').dropSchema;
+const dropIndexSchema = require('./schemas/collection_schemas').dropIndexSchema;
+const dropIndexesSchema = require('./schemas/collection_schemas').dropIndexesSchema;
+const estimatedDocumentCountSchema = require('./schemas/collection_schemas')
+  .estimatedDocumentCountSchema;
+const findSchema = require('./schemas/collection_schemas').findSchema;
+const findAndModifySchema = require('./schemas/collection_schemas').findAndModifySchema;
+const findAndRemoveSchema = require('./schemas/collection_schemas').findAndRemoveSchema;
+const findOneSchema = require('./schemas/collection_schemas').findOneSchema;
+const findOneAndDeleteSchema = require('./schemas/collection_schemas').findOneAndDeleteSchema;
+const findOneAndReplaceSchema = require('./schemas/collection_schemas').findOneAndReplaceSchema;
+const findOneAndUpdateSchema = require('./schemas/collection_schemas').findOneAndUpdateSchema;
+const geoHaystackSearchSchema = require('./schemas/collection_schemas').geoHaystackSearchSchema;
+const indexesSchema = require('./schemas/collection_schemas').indexesSchema;
+const indexExistsSchema = require('./schemas/collection_schemas').indexExistsSchema;
+const indexInformationSchema = require('./schemas/collection_schemas').indexInformationSchema;
+const initializeOrderedBulkOpSchema = require('./schemas/collection_schemas')
+  .initializeOrderedBulkOpSchema;
+const initializeUnorderedBulkOpSchema = require('./schemas/collection_schemas')
+  .initializeUnorderedBulkOpSchema;
+const insertManySchema = require('./schemas/collection_schemas').insertManySchema;
+const insertOneSchema = require('./schemas/collection_schemas').insertOneSchema;
+const isCappedSchema = require('./schemas/collection_schemas').isCappedSchema;
+const listIndexesSchema = require('./schemas/collection_schemas').listIndexesSchema;
+const mapReduceSchema = require('./schemas/collection_schemas').mapReduceSchema;
+const optionsSchema = require('./schemas/collection_schemas').optionsSchema;
+const parallelCollectionScanSchema = require('./schemas/collection_schemas')
+  .parallelCollectionScanSchema;
+const reIndexSchema = require('./schemas/collection_schemas').reIndexSchema;
+const renameSchema = require('./schemas/collection_schemas').renameSchema;
+const replaceOneSchema = require('./schemas/collection_schemas').replaceOneSchema;
+const statsSchema = require('./schemas/collection_schemas').statsSchema;
+const updateManySchema = require('./schemas/collection_schemas').updateManySchema;
+const updateOneSchema = require('./schemas/collection_schemas').updateOneSchema;
+const watchSchema = require('./schemas/collection_schemas').watchSchema;
 
 /**
  * @fileOverview The **Collection** class is an internal class that embodies a MongoDB collection
@@ -288,42 +332,6 @@ const DEPRECATED_FIND_OPTIONS = ['maxScan', 'fields', 'snapshot'];
  * @throws {MongoError}
  * @return {Cursor}
  */
-const findSchema = {
-  limit: { type: 'number', default: 0 },
-  sort: { type: ['array', 'object', 'string'] },
-  projection: { type: 'object' },
-  fields: { type: 'object' },
-  skip: { type: 'number', default: 0 },
-  hint: { type: ['string', 'object'] },
-  explain: { type: 'boolean' },
-  snapshot: { type: 'boolean' },
-  timeout: { type: 'boolean' },
-  tailable: { type: 'boolean' },
-  batchSize: { type: 'number' },
-  returnKey: { type: 'boolean' },
-  maxScan: { type: 'number' },
-  min: { type: 'number' },
-  max: { type: 'number' },
-  showDiskLoc: { type: 'boolean' },
-  showRecordId: { type: 'boolean' },
-  comment: { type: 'string' },
-  raw: { type: 'boolean' },
-  promoteLongs: { type: 'boolean' },
-  promoteValues: { type: 'boolean' },
-  promoteBuffers: { type: 'boolean' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  partial: { type: 'boolean' },
-  maxTimeMS: { type: 'number' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  promiseLibrary: { overrideOnly: true },
-  slaveOk: { type: 'boolean' },
-  awaitData: { type: 'boolean' },
-  noCursorTimeout: { type: 'boolean', overrideOnly: true },
-  ignoreUndefined: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.find = deprecateOptions(
   {
     name: 'collection.find',
@@ -419,16 +427,6 @@ Collection.prototype.find = deprecateOptions(
  * @param {Collection~insertOneWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const insertOneSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  serializeFunctions: { type: 'boolean' },
-  forceServerObjectId: { type: 'boolean' },
-  bypassDocumentValidation: { type: 'boolean' },
-  session: { type: ClientSession },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true }
-};
 Collection.prototype.insertOne = function(doc, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -478,17 +476,6 @@ function mapInsertManyResults(docs, r) {
  * @param {Collection~insertWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const insertManySchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  serializeFunctions: { type: 'boolean' },
-  forceServerObjectId: { type: 'boolean' },
-  bypassDocumentValidation: { type: 'boolean' },
-  ordered: { type: 'boolean', default: true },
-  session: { type: ClientSession },
-  ignoreUndefined: { overrideOnly: true }
-};
 Collection.prototype.insertMany = function(docs, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options ? Object.assign({}, options) : { ordered: true };
@@ -581,16 +568,6 @@ Collection.prototype.insertMany = function(docs, options, callback) {
  * @param {Collection~bulkWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const bulkWriteSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  serializeFunctions: { type: 'boolean' },
-  ordered: { type: 'boolean', default: true },
-  bypassDocumentValidation: { type: 'boolean' },
-  ignoreUndefined: { overrideOnly: true },
-  session: { type: ClientSession }
-};
 Collection.prototype.bulkWrite = function(operations, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -731,18 +708,6 @@ Collection.prototype.insert = deprecate(function(docs, options, callback) {
  * @param {Collection~updateWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const updateOneSchema = {
-  upsert: { type: 'boolean' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  bypassDocumentValidation: { type: 'boolean' },
-  arrayFilters: { type: 'array' },
-  collation: { type: 'object' },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true },
-  session: { type: ClientSession },
-  multi: { overrideOnly: true }
-};
 Collection.prototype.updateOne = function(filter, update, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -783,17 +748,6 @@ Collection.prototype.updateOne = function(filter, update, options, callback) {
  * @param {Collection~updateWriteOpCallback} [callback] The command result callback
  * @return {Promise<Collection~updatewriteOpResultObject>} returns Promise if no callback passed
  */
-const replaceOneSchema = {
-  upsert: { type: 'boolean' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  bypassDocumentValidation: { type: 'boolean' },
-  collation: { type: 'object' },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true },
-  session: { type: ClientSession },
-  multi: { overrideOnly: true }
-};
 Collection.prototype.replaceOne = function(filter, doc, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -829,18 +783,6 @@ Collection.prototype.replaceOne = function(filter, doc, options, callback) {
  * @param {Collection~updateWriteOpCallback} [callback] The command result callback
  * @return {Promise<Collection~updateWriteOpResultObject>} returns Promise if no callback passed
  */
-const updateManySchema = {
-  upsert: { type: 'boolean' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  arrayFilters: { type: 'array' },
-  bypassDocumentValidation: { type: 'boolean' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true },
-  multi: { overrideOnly: true }
-};
 Collection.prototype.updateMany = function(filter, update, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -933,15 +875,6 @@ Collection.prototype.update = deprecate(function(selector, update, options, call
  * @param {Collection~deleteWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const deleteOneSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true },
-  single: { overrideOnly: true }
-};
 Collection.prototype.deleteOne = function(filter, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -975,15 +908,6 @@ Collection.prototype.removeOne = Collection.prototype.deleteOne;
  * @param {Collection~deleteWriteOpCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const deleteManySchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  ignoreUndefined: { type: 'boolean', overrideOnly: true },
-  single: { overrideOnly: true }
-};
 Collection.prototype.deleteMany = function(filter, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1104,32 +1028,6 @@ Collection.prototype.save = deprecate(function(doc, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const findOneSchema = {
-  projection: { type: 'object' },
-  fields: { type: 'object' },
-  hint: { type: ['string', 'object'] },
-  explain: { type: 'boolean' },
-  snapshot: { type: 'boolean' },
-  timeout: { type: 'boolean' },
-  tailable: { type: 'boolean' },
-  batchSize: { type: 'number' },
-  returnKey: { type: 'boolean' },
-  maxScan: { type: 'number' },
-  min: { type: 'number' },
-  max: { type: 'number' },
-  showDiskLoc: { type: 'boolean' },
-  showRecordId: { type: 'boolean' },
-  comment: { type: 'string' },
-  raw: { type: 'boolean' },
-  promoteLongs: { type: 'boolean' },
-  promoteValues: { type: 'boolean' },
-  promoteBuffers: { type: 'boolean' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  partial: { type: 'boolean' },
-  maxTimeMS: { type: 'number' },
-  collation: { type: 'object' },
-  session: { type: ClientSession }
-};
 Collection.prototype.findOne = deprecateOptions(
   {
     name: 'collection.find',
@@ -1172,11 +1070,6 @@ Collection.prototype.findOne = deprecateOptions(
  * @param {Collection~collectionResultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const renameSchema = {
-  dropTarget: { type: 'boolean', default: false },
-  readPreference: { type: [ReadPreference, 'string', 'object'], overrideOnly: true },
-  session: { type: ClientSession }
-};
 Collection.prototype.rename = function(newName, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1199,13 +1092,6 @@ Collection.prototype.rename = function(newName, options, callback) {
  * @param {Collection~resultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const dropSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  readPreference: { type: [ReadPreference, 'string', 'object'], overrideOnly: true },
-  session: { type: ClientSession }
-};
 Collection.prototype.drop = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1231,9 +1117,6 @@ Collection.prototype.drop = function(options, callback) {
  * @param {Collection~resultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const optionsSchema = {
-  session: { type: ClientSession }
-};
 Collection.prototype.options = function(opts, callback) {
   if (typeof opts === 'function') (callback = opts), (opts = {});
 
@@ -1253,9 +1136,6 @@ Collection.prototype.options = function(opts, callback) {
  * @param {Collection~resultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const isCappedSchema = {
-  session: { type: ClientSession }
-};
 Collection.prototype.isCapped = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1289,24 +1169,6 @@ Collection.prototype.isCapped = function(options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const createIndexSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  unique: { type: 'boolean', default: false },
-  sparse: { type: 'boolean' },
-  background: { type: 'boolean' },
-  dropDups: { type: 'boolean' },
-  min: { type: 'number' },
-  max: { type: 'number' },
-  v: { type: 'number' },
-  expireAfterSeconds: { type: 'number' },
-  name: { type: 'string' },
-  partialFilterExpression: { type: 'object' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  readPreference: { overrideOnly: true }
-};
 Collection.prototype.createIndex = function(fieldOrSpec, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1337,14 +1199,6 @@ Collection.prototype.createIndex = function(fieldOrSpec, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const createIndexesSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  maxTimeMS: { type: 'number' },
-  readPreference: { overrideOnly: true },
-  session: { type: ClientSession }
-};
 Collection.prototype.createIndexes = function(indexSpecs, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1373,14 +1227,6 @@ Collection.prototype.createIndexes = function(indexSpecs, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const dropIndexSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  readPreference: { type: [ReadPreference, 'string', 'object'], overrideOnly: true },
-  maxTimeMS: { type: 'number' },
-  session: { type: ClientSession }
-};
 Collection.prototype.dropIndex = function(indexName, options, callback) {
   const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -1405,13 +1251,6 @@ Collection.prototype.dropIndex = function(indexName, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const dropIndexesSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  maxTimeMS: { type: 'number' },
-  session: { type: ClientSession }
-};
 Collection.prototype.dropIndexes = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1443,9 +1282,6 @@ Collection.prototype.dropAllIndexes = deprecate(
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const reIndexSchema = {
-  session: { type: ClientSession }
-};
 Collection.prototype.reIndex = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1466,14 +1302,6 @@ Collection.prototype.reIndex = function(options, callback) {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {CommandCursor}
  */
-const listIndexesSchema = {
-  batchSize: { type: 'number' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession },
-  cursorFactory: { type: 'function', overrideOnly: true },
-  promiseLibrary: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true }
-};
 Collection.prototype.listIndexes = function(options) {
   if (!this.s.topology.capabilities()) {
     throw new MongoError('cannot connect to server');
@@ -1535,10 +1363,6 @@ Collection.prototype.ensureIndex = deprecate(function(fieldOrSpec, options, call
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const indexExistsSchema = {
-  session: { type: ClientSession },
-  full: { type: 'boolean', default: false }
-};
 Collection.prototype.indexExists = function(indexes, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1558,10 +1382,6 @@ Collection.prototype.indexExists = function(indexes, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const indexInformationSchema = {
-  full: { type: 'boolean', default: false },
-  session: { type: ClientSession }
-};
 Collection.prototype.indexInformation = function(options, callback) {
   const args = Array.prototype.slice.call(arguments, 0);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -1617,10 +1437,6 @@ Collection.prototype.count = deprecate(function(query, options, callback) {
  * @param {Collection~countCallback} [callback] The command result callback.
  * @return {Promise} returns Promise if no callback passed.
  */
-const estimatedDocumentCountSchema = {
-  maxTimeMS: { type: 'number' },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.estimatedDocumentCount = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1665,14 +1481,6 @@ Collection.prototype.estimatedDocumentCount = function(options, callback) {
  * @see https://docs.mongodb.com/manual/reference/operator/query/center/#op._S_center
  * @see https://docs.mongodb.com/manual/reference/operator/query/centerSphere/#op._S_centerSphere
  */
-const countDocumentsSchema = {
-  collation: { type: 'object' },
-  hint: { type: ['string', 'object'] },
-  limit: { type: 'number' },
-  maxTimeMS: { type: 'number' },
-  skip: { type: 'number' },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.countDocuments = function(query, options, callback) {
   const args = Array.prototype.slice.call(arguments, 0);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -1700,13 +1508,6 @@ Collection.prototype.countDocuments = function(query, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const distinctSchema = {
-  collation: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'string', 'object'] },
-  maxTimeMS: { type: 'number' },
-  session: { type: ClientSession },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.distinct = function(key, query, options, callback) {
   const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -1735,12 +1536,6 @@ Collection.prototype.distinct = function(key, query, options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const indexesSchema = {
-  batchSize: { type: 'number' },
-  readPreference: { type: [ReadPreference, 'string', 'object'] },
-  session: { type: ClientSession },
-  full: { type: 'boolean', default: true }
-};
 Collection.prototype.indexes = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1762,11 +1557,6 @@ Collection.prototype.indexes = function(options, callback) {
  * @param {Collection~resultCallback} [callback] The collection result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const statsSchema = {
-  readPreference: { type: [ReadPreference, 'string', 'object'] },
-  scale: { type: 'number' },
-  session: { type: ClientSession }
-};
 Collection.prototype.stats = function(options, callback) {
   const args = Array.prototype.slice.call(arguments, 0);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -1811,21 +1601,6 @@ Collection.prototype.stats = function(options, callback) {
  * @param {Collection~findAndModifyCallback} [callback] The collection result callback
  * @return {Promise<Collection~findAndModifyWriteOpResultObject>} returns Promise if no callback passed
  */
-const findOneAndDeleteSchema = {
-  collation: { type: 'object' },
-  projection: { type: 'object' },
-  sort: { type: ['object', 'string'] },
-  maxTimeMS: { type: 'number' },
-  session: { type: ClientSession },
-  fields: { overrideOnly: true },
-  remove: { overrideOnly: true },
-  new: { overrideOnly: true },
-  upsert: { overrideOnly: true },
-  serializeFunctions: { type: 'boolean' },
-  checkKeys: { overrideOnly: true },
-  readPreference: { overrideOnly: true },
-  fsync: { type: 'number' }
-};
 Collection.prototype.findOneAndDelete = function(filter, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1872,24 +1647,6 @@ Collection.prototype.findOneAndDelete = function(filter, options, callback) {
  * @param {Collection~findAndModifyCallback} [callback] The collection result callback
  * @return {Promise<Collection~findAndModifyWriteOpResultObject>} returns Promise if no callback passed
  */
-const findOneAndReplaceSchema = {
-  bypassDocumentValidation: { type: 'boolean' },
-  collation: { type: 'object' },
-  projection: { type: 'object' },
-  sort: { type: ['object', 'string'] },
-  maxTimeMS: { type: 'number' },
-  upsert: { type: 'boolean', default: false },
-  returnOriginal: { type: 'boolean' },
-  session: { type: ClientSession },
-  fields: { overrideOnly: true },
-  update: { overrideOnly: true },
-  new: { overrideOnly: true },
-  remove: { overrideOnly: true },
-  serializeFunctions: { type: 'boolean' },
-  checkKeys: { overrideOnly: true },
-  readPreference: { overrideOnly: true },
-  fsync: { type: 'number' }
-};
 Collection.prototype.findOneAndReplace = function(filter, replacement, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1945,25 +1702,6 @@ Collection.prototype.findOneAndReplace = function(filter, replacement, options, 
  * @param {Collection~findAndModifyCallback} [callback] The collection result callback
  * @return {Promise<Collection~findAndModifyWriteOpResultObject>} returns Promise if no callback passed
  */
-const findOneAndUpdateSchema = {
-  bypassDocumentValidation: { type: 'boolean' },
-  collation: { type: 'object' },
-  projection: { type: 'object' },
-  sort: { type: ['object', 'string'] },
-  maxTimeMS: { type: 'number' },
-  upsert: { type: 'boolean', default: false },
-  returnOriginal: { type: 'boolean' },
-  session: { type: ClientSession },
-  arrayFilters: { type: 'array' },
-  fields: { overrideOnly: true },
-  update: { overrideOnly: true },
-  new: { overrideOnly: true },
-  remove: { overrideOnly: true },
-  serializeFunctions: { type: 'boolean' },
-  checkKeys: { overrideOnly: true },
-  readPreference: { overrideOnly: true },
-  fsync: { type: 'number' }
-};
 Collection.prototype.findOneAndUpdate = function(filter, update, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -2027,22 +1765,6 @@ Collection.prototype.findOneAndUpdate = function(filter, update, options, callba
  * @return {Promise} returns Promise if no callback passed
  * @deprecated use findOneAndUpdate, findOneAndReplace or findOneAndDelete instead
  */
-const findAndModifySchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  remove: { type: 'boolean', default: false },
-  upsert: { type: 'boolean', default: false },
-  new: { type: 'boolean', default: false },
-  projection: { type: 'object' },
-  fields: { type: 'object' },
-  session: { type: ClientSession },
-  arrayFilters: { type: 'array' },
-  readPreference: { overrideOnly: true },
-  serializeFunctions: { type: 'boolean' },
-  checkKeys: { overrideOnly: true },
-  collation: { type: 'object' }
-};
 Collection.prototype.findAndModify = deprecate(function(query, sort, doc, options, callback) {
   const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -2084,18 +1806,6 @@ Collection.prototype.findAndModify = deprecate(function(query, sort, doc, option
  * @return {Promise} returns Promise if no callback passed
  * @deprecated use findOneAndDelete instead
  */
-const findAndRemoveSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession },
-  remove: { overrideOnly: true },
-  readPreference: { overrideOnly: true },
-  serializeFunctions: { type: 'boolean' },
-  checkKeys: { overrideOnly: true },
-  upsert: { overrideOnly: true },
-  new: { overrideOnly: true }
-};
 Collection.prototype.findAndRemove = deprecate(function(query, sort, options, callback) {
   const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -2144,30 +1854,6 @@ Collection.prototype.findAndRemove = deprecate(function(query, sort, options, ca
  * @param {Collection~aggregationCallback} callback The command result callback
  * @return {(null|AggregationCursor)}
  */
-const aggregateSchema = {
-  readPreference: { type: [ReadPreference, 'string', 'object'] },
-  cursor: { type: 'object', default: {} },
-  batchSize: { type: 'number' },
-  explain: { type: 'boolean' },
-  allowDiskUse: { type: 'boolean' },
-  maxTimeMS: { type: 'number' },
-  bypassDocumentValidation: { type: 'boolean' },
-  raw: { type: 'boolean' },
-  promoteLongs: { type: 'boolean' },
-  promoteValues: { type: 'boolean' },
-  promoteBuffers: { type: 'boolean' },
-  collation: { type: 'object' },
-  comment: { type: 'string' },
-  hint: { type: ['object', 'string'] },
-  session: { type: ClientSession },
-  promiseLibrary: { type: 'function', overrideOnly: true },
-  cursorFactory: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true },
-  readConcern: { type: 'object' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' }
-};
 Collection.prototype.aggregate = function(pipeline, options, callback) {
   if (Array.isArray(pipeline)) {
     // Set up callback if one is provided
@@ -2237,17 +1923,6 @@ Collection.prototype.aggregate = function(pipeline, options, callback) {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {ChangeStream} a ChangeStream instance.
  */
-const watchSchema = {
-  fullDocument: { type: 'string', default: 'default' },
-  resumeAfter: { type: 'object' },
-  maxAwaitTimeMS: { type: 'number' },
-  batchSize: { type: 'number' },
-  collation: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  startAtClusterTime: { type: 'Timestamp' },
-  session: { type: ClientSession },
-  optionsValidationLevel: { overrideOnly: true }
-};
 Collection.prototype.watch = function(pipeline, options) {
   pipeline = pipeline || [];
 
@@ -2293,16 +1968,6 @@ Collection.prototype.watch = function(pipeline, options) {
  * @param {Collection~parallelCollectionScanCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const parallelCollectionScanSchema = {
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  batchSize: { type: 'number', default: 1000 },
-  numCursors: { type: 'number', default: 1 },
-  raw: { type: 'boolean', default: false },
-  promiseLibrary: { overrideOnly: true },
-  session: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.parallelCollectionScan = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = { numCursors: 1 });
 
@@ -2343,14 +2008,6 @@ Collection.prototype.parallelCollectionScan = function(options, callback) {
  * @param {Collection~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const geoHaystackSearchSchema = {
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  maxDistance: { type: 'number' },
-  search: { type: 'object' },
-  limit: { type: 'number' },
-  session: { type: ClientSession },
-  readConcern: { type: 'object' }
-};
 Collection.prototype.geoHaystackSearch = function(x, y, options, callback) {
   const args = Array.prototype.slice.call(arguments, 2);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
@@ -2469,25 +2126,6 @@ Collection.prototype.group = deprecate(function(
  * @throws {MongoError}
  * @return {Promise} returns Promise if no callback passed
  */
-const mapReduceSchema = {
-  collation: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'string', 'object'] },
-  out: { type: ['string', 'object'], required: true },
-  query: { type: 'object' },
-  sort: { type: 'object' },
-  limit: { type: 'number' },
-  keeptemp: { type: 'boolean' },
-  finalize: { type: ['function', 'string'] },
-  scope: { type: 'object' },
-  jsMode: { type: 'boolean' },
-  verbose: { type: 'boolean' },
-  bypassDocumentValidation: { type: 'boolean' },
-  session: { type: ClientSession },
-  readConcern: { type: 'object' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' }
-};
 Collection.prototype.mapReduce = function(map, reduce, options, callback) {
   if ('function' === typeof options) (callback = options), (options = {});
 
@@ -2531,15 +2169,6 @@ Collection.prototype.mapReduce = function(map, reduce, options, callback) {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {UnorderedBulkOperation}
  */
-const initializeUnorderedBulkOpSchema = {
-  bypassDocumentValidation: { type: 'boolean' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession },
-  promiseLibrary: { overrideOnly: true },
-  ignoreUndefined: { type: 'boolean' }
-};
 Collection.prototype.initializeUnorderedBulkOp = function(options) {
   validate(initializeUnorderedBulkOpSchema, options, {
     optionsValidationLevel: this.optionsValidationLevel
@@ -2568,15 +2197,6 @@ Collection.prototype.initializeUnorderedBulkOp = function(options) {
  * @param {OrderedBulkOperation} callback The command result callback
  * @return {null}
  */
-const initializeOrderedBulkOpSchema = {
-  bypassDocumentValidation: { type: 'boolean' },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession },
-  promiseLibrary: { overrideOnly: true },
-  ignoreUndefined: { type: 'boolean' }
-};
 Collection.prototype.initializeOrderedBulkOp = function(options) {
   validate(initializeOrderedBulkOpSchema, options, {
     optionsValidationLevel: this.optionsValidationLevel
