+++ lib/db.js
@@ -14,7 +14,6 @@ const applyWriteConcern = require('./utils').applyWriteConcern;
 const resolveReadPreference = require('./utils').resolveReadPreference;
 const ChangeStream = require('./change_stream');
 const deprecate = require('util').deprecate;
-const ClientSession = require('mongodb-core').Sessions.ClientSession;
 const validate = require('./options_validator').validate;
 const applyDefaults = require('./options_validator').applyDefaults;
 const CONSTANTS = require('./constants');
@@ -40,6 +39,25 @@ const removeUser = require('./operations/db_ops').removeUser;
 const rename = require('./operations/collection_ops').rename;
 const setProfilingLevel = require('./operations/db_ops').setProfilingLevel;
 const validateDatabaseName = require('./operations/db_ops').validateDatabaseName;
+
+// Schemas
+const addUserSchema = require('./schemas/db_schemas').addUserSchema;
+const collectionSchema = require('./schemas/db_schemas').collectionSchema;
+const collectionsSchema = require('./schemas/db_schemas').collectionsSchema;
+const commandSchema = require('./schemas/db_schemas').commandSchema;
+const createCollectionSchema = require('./schemas/db_schemas').createCollectionSchema;
+const createIndexSchema = require('./schemas/db_schemas').createIndexSchema;
+const dropCollectionSchema = require('./schemas/db_schemas').dropCollectionSchema;
+const dropDatabaseSchema = require('./schemas/db_schemas').dropDatabaseSchema;
+const executeDbAdminCommandSchema = require('./schemas/db_schemas').executeDbAdminCommandSchema;
+const indexInformationSchema = require('./schemas/db_schemas').indexInformationSchema;
+const listCollectionsSchema = require('./schemas/db_schemas').listCollectionsSchema;
+const profilingLevelSchema = require('./schemas/db_schemas').profilingLevelSchema;
+const removeUserSchema = require('./schemas/db_schemas').removeUserSchema;
+const renameCollectionSchema = require('./schemas/db_schemas').renameCollectionSchema;
+const setProfilingLevelSchema = require('./schemas/db_schemas').setProfilingLevelSchema;
+const statsSchema = require('./schemas/db_schemas').statsSchema;
+const watchSchema = require('./schemas/db_schemas').watchSchema;
 
 /**
  * @fileOverview The **Db** class is a class that represents a MongoDB Database.
@@ -282,10 +300,6 @@ Object.defineProperty(Db.prototype, 'writeConcern', {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const commandSchema = {
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession }
-};
 Db.prototype.command = function(command, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -333,20 +347,6 @@ Db.prototype.admin = function() {
  * @param {Db~collectionResultCallback} [callback] The collection result callback
  * @return {Collection} return the new Collection instance if not in strict mode
  */
-const collectionSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  raw: { type: 'boolean' },
-  pkFactory: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  serializeFunctions: { type: 'boolean' },
-  strict: { type: 'boolean' },
-  readConcern: { type: 'object' },
-  promiseLibrary: { overrideOnly: true },
-  ignoreUndefined: { overrideOnly: true },
-  fsync: { type: 'number' }
-};
 Db.prototype.collection = function(name, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -399,31 +399,6 @@ Db.prototype.collection = function(name, options, callback) {
  * @param {Db~collectionResultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const createCollectionSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  raw: { type: 'boolean' },
-  pkFactory: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  serializeFunctions: { type: 'boolean' },
-  strict: { type: 'boolean' },
-  capped: { type: 'boolean' },
-  autoIndexId: { type: 'boolean', deprecated: true },
-  size: { type: 'number' },
-  max: { type: 'number' },
-  flags: { type: 'number' },
-  storageEngine: { type: 'object' },
-  validator: { type: 'object' },
-  validationLevel: { type: 'string' },
-  validationAction: { type: 'string' },
-  indexOptionDefaults: { type: 'object' },
-  viewOn: { type: 'string' },
-  pipeline: { type: 'array' },
-  collation: { type: 'object' },
-  session: { type: ClientSession },
-  promiseLibrary: { type: 'function' }
-};
 Db.prototype.createCollection = function(name, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -449,11 +424,6 @@ Db.prototype.createCollection = function(name, options, callback) {
  * @param {Db~resultCallback} [callback] The collection result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const statsSchema = {
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  scale: { type: 'number' },
-  session: { type: ClientSession }
-};
 Db.prototype.stats = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -486,14 +456,6 @@ Db.prototype.stats = function(options, callback) {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {CommandCursor}
  */
-const listCollectionsSchema = {
-  nameOnly: { type: 'boolean', default: false },
-  batchSize: { type: 'number' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession },
-  cursorFactory: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true }
-};
 Db.prototype.listCollections = function(filter, options) {
   filter = filter || {};
 
@@ -551,14 +513,6 @@ Db.prototype.eval = deprecate(function(code, parameters, options, callback) {
  * @param {Db~collectionResultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const renameCollectionSchema = {
-  dropTarget: { type: 'boolean', default: false },
-  session: { type: ClientSession },
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  readPreference: { type: [ReadPreference, 'string', 'object'], overrideOnly: true }
-};
 Db.prototype.renameCollection = function(fromCollection, toCollection, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -593,13 +547,6 @@ Db.prototype.renameCollection = function(fromCollection, toCollection, options, 
  * @param {Db~resultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const dropCollectionSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession },
-  readPreference: { overrideOnly: true }
-};
 Db.prototype.dropCollection = function(name, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -631,14 +578,6 @@ Db.prototype.dropCollection = function(name, options, callback) {
  * @param {Db~resultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const dropDatabaseSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession },
-  readPreference: { overrideOnly: true },
-  writeConcern: { type: 'object' }
-};
 Db.prototype.dropDatabase = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -671,12 +610,6 @@ Db.prototype.dropDatabase = function(options, callback) {
  * @param {Db~collectionsResultCallback} [callback] The results callback
  * @return {Promise} returns Promise if no callback passed
  */
-const collectionsSchema = {
-  nameOnly: { type: 'boolean', default: false },
-  batchSize: { type: 'number' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession }
-};
 Db.prototype.collections = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -697,10 +630,6 @@ Db.prototype.collections = function(options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const executeDbAdminCommandSchema = {
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession }
-};
 Db.prototype.executeDbAdminCommand = function(selector, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -747,23 +676,6 @@ Db.prototype.executeDbAdminCommand = function(selector, options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const createIndexSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  unique: { type: 'boolean', default: false },
-  sparse: { type: 'boolean' },
-  background: { type: 'boolean' },
-  dropDups: { type: 'boolean' },
-  min: { type: 'number' },
-  max: { type: 'number' },
-  v: { type: 'number' },
-  expireAfterSeconds: { type: 'number' },
-  name: { type: 'number' },
-  partialFilterExpression: { type: 'object' },
-  collation: { type: 'object' },
-  session: { type: ClientSession }
-};
 Db.prototype.createIndex = function(name, fieldOrSpec, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -839,17 +751,6 @@ Db.prototype.addChild = function(db) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const addUserSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  customData: { type: 'object' },
-  roles: { type: 'object' },
-  authenticationRestrictions: { type: 'array' },
-  mechanisms: { type: 'array' },
-  passwordDigestor: { type: 'string' },
-  session: { type: ClientSession }
-};
 Db.prototype.addUser = function(username, password, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -872,12 +773,6 @@ Db.prototype.addUser = function(username, password, options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const removeUserSchema = {
-  w: { type: ['number', 'string'] },
-  wtimeout: { type: 'number' },
-  j: { type: 'boolean' },
-  session: { type: ClientSession }
-};
 Db.prototype.removeUser = function(username, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -897,9 +792,6 @@ Db.prototype.removeUser = function(username, options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback.
  * @return {Promise} returns Promise if no callback passed
  */
-const setProfilingLevelSchema = {
-  session: { type: ClientSession }
-};
 Db.prototype.setProfilingLevel = function(level, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -936,9 +828,6 @@ Db.prototype.profilingInfo = deprecate(function(options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const profilingLevelSchema = {
-  session: { type: ClientSession }
-};
 Db.prototype.profilingLevel = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -962,12 +851,6 @@ Db.prototype.profilingLevel = function(options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const indexInformationSchema = {
-  batchSize: { type: 'number' },
-  full: { type: 'boolean', default: false },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession }
-};
 Db.prototype.indexInformation = function(name, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -1004,17 +887,6 @@ Db.prototype.unref = function() {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {ChangeStream} a ChangeStream instance.
  */
-const watchSchema = {
-  fullDocument: { type: 'string', default: 'default' },
-  resumeAfter: { type: 'object' },
-  maxAwaitTimeMS: { type: 'number' },
-  batchSize: { type: 'number' },
-  collation: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  startAtClusterTime: { type: 'Timestamp' },
-  session: { type: ClientSession },
-  optionsValidationLevel: { overrideOnly: true }
-};
 Db.prototype.watch = function(pipeline, options) {
   pipeline = pipeline || [];
