+++ lib/mongo_client.js
@@ -2,13 +2,11 @@
 
 const applyDefaults = require('./options_validator').applyDefaults;
 const ChangeStream = require('./change_stream');
-const ClientSession = require('mongodb-core').Sessions.ClientSession;
 const Db = require('./db');
 const EventEmitter = require('events').EventEmitter;
 const executeOperation = require('./utils').executeOperation;
 const inherits = require('util').inherits;
 const MongoError = require('mongodb-core').MongoError;
-const ReadPreference = require('mongodb-core').ReadPreference;
 const resolveReadPreference = require('./utils').resolveReadPreference;
 const validate = require('./options_validator').validate;
 
@@ -16,89 +14,14 @@ const validate = require('./options_validator').validate;
 const connectOp = require('./operations/mongo_client_ops').connectOp;
 const logout = require('./operations/mongo_client_ops').logout;
 
-const connectSchema = {
-  appname: { type: 'string' },
-  authMechanism: { type: 'string' },
-  authMechanismProperties: { type: 'array' },
-  authSource: { type: 'string' },
-  compression: { type: 'string', alias: 'compressors' },
-  connectTimeoutMS: { type: 'number', default: 30000 },
-  heartbeatFrequencyMS: { type: 'number', default: 60000 },
-  j: { type: 'boolean', alias: 'journal' },
-  secondaryAcceptableLatencyMS: { type: 'number', default: 15, alias: 'localThresholdMS' },
-  maxIdleTimeMS: { type: 'number' },
-  maxPoolSize: { type: 'number', default: 5, alias: 'poolSize' },
-  poolSize: { type: 'number' },
-  maxStalenessSeconds: { type: 'number' },
-  readConcernLevel: { type: 'string', alias: 'readConcern.level' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  readPreferenceTags: { type: 'object' },
-  replicaSet: { type: 'string' },
-  retryWrites: { type: 'boolean', default: false },
-  serverSelectionTimeoutMS: { type: 'number', default: 30000 },
-  serverSelectionTryOnce: { type: 'boolean', default: true },
-  socketTimeoutMS: { type: 'number', default: 360000 },
-  ssl: { type: 'boolean', default: false },
-  tls: { type: 'boolean', default: false },
-  tlsAllowInvalidCertificates: { type: 'boolean', default: false },
-  tlsAllowInvalidHostnames: { type: 'boolean', default: false },
-  tlsCAFile: { type: 'string' },
-  tlsCertificateKeyFile: { type: 'string' },
-  tlsCertificateKeyFilePassword: { type: 'string' },
-  tlsInsecure: { type: 'boolean', default: false },
-  w: { type: 'number' },
-  wtimeout: { type: 'number', alias: 'wTimeoutMS' },
-  zlibCompressionLevel: { type: 'number', default: -1 },
-  // TODO: check these old options
-  sslCA: { type: 'string' },
-  sslCert: { type: 'string' },
-  sslValidate: { type: 'boolean', default: true },
-  sslKey: { type: 'buffer' },
-  sslPass: { type: 'string' },
-  sslCRL: { type: 'buffer' },
-  autoReconnect: { type: 'boolean', default: true },
-  auto_reconnect: { type: 'boolean', default: true },
-  noDelay: { type: 'boolean', default: true },
-  keepAlive: { type: 'boolean', default: true },
-  keepAliveInitialDelay: { type: 'number', default: 30000 },
-  family: { type: 'number' },
-  reconnectTries: { type: 'number', default: 30 },
-  reconnectInterval: { type: 'number', default: 1000 },
-  ha: { type: 'boolean', default: true },
-  haInterval: { type: 'number', default: 10000 },
-  acceptableLatencyMS: { type: 'number', default: 15 },
-  connectWithNoPrimary: { type: 'boolean', default: false },
-  forceServerObjectId: { type: 'boolean', default: false },
-  serializeFunctions: { type: 'boolean', default: false },
-  ignoreUndefined: { type: 'boolean', default: false },
-  raw: { type: 'boolean', default: false },
-  bufferMaxEntries: { type: 'number', default: -1 },
-  pkFactory: { type: ['object', 'function'] },
-  promiseLibrary: { type: ['object', 'function'] },
-  readConcern: { type: 'object' },
-  'readConcern.level': { type: 'string' },
-  loggerLevel: { type: 'string' },
-  logger: { type: ['object', 'function'] },
-  promoteValues: { type: 'boolean' },
-  promoteBuffers: { type: 'boolean' },
-  promoteLongs: { type: 'boolean' },
-  domainsEnabled: { type: 'boolean', default: false },
-  validateOptions: { type: 'boolean', default: false, deprecated: 'unknownOptionsWarningLevel' },
-  checkServerIdentity: { type: ['boolean', 'function'], default: true },
-  'auth.user': { type: 'string' },
-  'auth.password': { type: 'string' },
-  fsync: { type: 'boolean' },
-  numberOfRetries: { type: 'number', default: 5 },
-  minSize: { type: 'number' },
-  emitError: { type: 'boolean', default: true },
-  monitorCommands: { type: 'boolean', default: false },
-  useNewUrlParser: { type: 'boolean' },
-  useUnifiedTopology: { type: 'boolean' },
-  optionsValidationLevel: { type: 'string' },
-  unknownOptionsWarningLevel: { type: 'string' },
-  host: { type: 'string' },
-  auth: { type: 'object' }
-};
+// Schemas
+const connectSchema = require('./schemas/mongo_client_schemas').connectSchema;
+const dbSchema = require('./schemas/mongo_client_schemas').dbSchema;
+const isConnectedSchema = require('./schemas/mongo_client_schemas').isConnectedSchema;
+const logoutSchema = require('./schemas/mongo_client_schemas').logoutSchema;
+const startSessionSchema = require('./schemas/mongo_client_schemas').startSessionSchema;
+const watchSchema = require('./schemas/mongo_client_schemas').watchSchema;
+const withSessionSchema = require('./schemas/mongo_client_schemas').withSessionSchema;
 
 /**
  * @fileOverview The **MongoClient** class is a class that allows for making Connections to MongoDB.
@@ -340,9 +263,6 @@ MongoClient.prototype.connect = function(callback) {
  * @param {MongoClient~logoutCallback} [callback] The command result callback
  * @return {Promise} returns Promise if no callback passed
  */
-const logoutSchema = {
-  dbName: { type: 'string' }
-};
 MongoClient.prototype.logout = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
 
@@ -415,13 +335,6 @@ MongoClient.prototype.close = function(force, callback) {
  * @param {boolean} [options.returnNonCachedInstance] Control if you want to return a cached instance or have a new one created
  * @return {Db}
  */
-const dbSchema = {
-  noListener: { type: 'boolean', default: false },
-  returnNonCachedInstance: { type: 'boolean' },
-  readPreference: { overrideOnly: true },
-  promiseLibrary: { overrideOnly: true },
-  optionsValidationLevel: { type: 'string' }
-};
 MongoClient.prototype.db = function(dbName, options) {
   validate(dbSchema, options, { optionsValidationLevel: this.optionsValidationLevel });
 
@@ -471,10 +384,6 @@ MongoClient.prototype.db = function(dbName, options) {
  * @param {boolean} [options.returnNonCachedInstance] Control if you want to return a cached instance or have a new one created
  * @return {boolean}
  */
-const isConnectedSchema = {
-  noListener: { type: 'boolean', default: false },
-  returnNonCachedInstance: { type: 'boolean' }
-};
 MongoClient.prototype.isConnected = function(options) {
   validate(isConnectedSchema, options, { optionsValidationLevel: this.optionsValidationLevel });
 
@@ -611,10 +520,6 @@ MongoClient.connect = function(url, options, callback) {
  * @param {SessionOptions} [options] optional settings for a driver session
  * @return {ClientSession} the newly established session
  */
-const startSessionSchema = {
-  explicit: { overrideOnly: true },
-  causalConsistency: { type: 'boolean' }
-};
 MongoClient.prototype.startSession = function(options) {
   validate(startSessionSchema, options, { optionsValidationLevel: this.optionsValidationLevel });
 
@@ -641,9 +546,6 @@ MongoClient.prototype.startSession = function(options) {
  * @param {Function} operation An operation to execute with an implicitly created session. The signature of this MUST be `(session) => {}`
  * @return {Promise}
  */
-const withSessionSchema = {
-  causalConsistency: { type: 'boolean' }
-};
 MongoClient.prototype.withSession = function(options, operation) {
   if (typeof options === 'function') (operation = options), (options = undefined);
   const session = this.startSession(options);
@@ -693,17 +595,6 @@ MongoClient.prototype.withSession = function(options, operation) {
  * @param {ClientSession} [options.session] optional session to use for this operation
  * @return {ChangeStream} a ChangeStream instance.
  */
-const watchSchema = {
-  fullDocument: { type: 'string', default: 'default' },
-  resumeAfter: { type: 'object' },
-  maxAwaitTimeMS: { type: 'number' },
-  batchSize: { type: 'number' },
-  collation: { type: 'object' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  startAtClusterTime: { type: 'Timestamp' },
-  session: { type: ClientSession },
-  optionsValidationLevel: { overrideOnly: true }
-};
 MongoClient.prototype.watch = function(pipeline, options) {
   pipeline = pipeline || [];
