+++ lib/operations/collection_ops.js
@@ -23,10 +23,10 @@ const Long = require('mongodb-core').BSON.Long;
 const makeLazyLoader = require('../utils').makeLazyLoader;
 const MongoError = require('mongodb-core').MongoError;
 const normalizeHintField = require('../utils').normalizeHintField;
-const ordered = require('../bulk/ordered');
+const initializeOrderedBulkOp = require('../bulk/ordered');
 const ReadPreference = require('mongodb-core').ReadPreference;
 const toError = require('../utils').toError;
-const unordered = require('../bulk/unordered');
+const initializeUnorderedBulkOp = require('../bulk/unordered');
 
 const loadCollection = makeLazyLoader(`${__dirname}/../collection`);
 const loadDb = makeLazyLoader(`${__dirname}/../db`);
@@ -190,8 +190,8 @@ function bulkWrite(coll, operations, options, callback) {
   // Create the bulk operation
   const bulk =
     options.ordered === true
-      ? ordered(coll.s.topology, coll, options)
-      : unordered(coll.s.topology, coll, options);
+      ? initializeOrderedBulkOp(coll.s.topology, coll, options)
+      : initializeUnorderedBulkOp(coll.s.topology, coll, options);
 
   // Do we have a collation
   let collation = false;
