+++ lib/operations/db_ops.js
@@ -2,13 +2,13 @@
 
 const applyDefaults = require('../options_validator').applyDefaults;
 const applyWriteConcern = require('../utils').applyWriteConcern;
-const ClientSession = require('mongodb-core').Sessions.ClientSession;
 const Code = require('mongodb-core').BSON.Code;
 const resolveReadPreference = require('../utils').resolveReadPreference;
 const crypto = require('crypto');
 const debugOptions = require('../utils').debugOptions;
 const handleCallback = require('../utils').handleCallback;
 const makeLazyLoader = require('../utils').makeLazyLoader;
+const mergeOptionsAndWriteConcern = require('../utils').mergeOptionsAndWriteConcern;
 const MongoError = require('mongodb-core').MongoError;
 const parseIndexOptions = require('../utils').parseIndexOptions;
 const ReadPreference = require('mongodb-core').ReadPreference;
@@ -20,8 +20,22 @@ const findOne = require('./collection_ops').findOne;
 const remove = require('./collection_ops').remove;
 const updateOne = require('./collection_ops').updateOne;
 
+const listCollectionsSchema = require('./schemas/collection_schemas').listCollectionsSchema;
+
 const loadCollection = makeLazyLoader(`${__dirname}/../collection`);
 const loadDb = makeLazyLoader(`${__dirname}/../db`);
+
+const collectionKeys = [
+  'pkFactory',
+  'readPreference',
+  'serializeFunctions',
+  'strict',
+  'readConcern',
+  'ignoreUndefined',
+  'promoteValues',
+  'promoteBuffers',
+  'promoteLongs'
+];
 
 const debugFields = [
   'authSource',
@@ -130,7 +144,7 @@ function addUser(db, username, password, options, callback) {
 }
 
 function collection(db, name, options, callback) {
-  const Collection = require('../collection');
+  let Collection = loadCollection();
 
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
@@ -319,7 +333,7 @@ function createCollection(db, name, options, callback) {
  * @param {Db~resultCallback} [callback] The command result callback
  */
 function createIndex(db, name, fieldOrSpec, options, callback) {
-  const Db = require('../db');
+  let Db = loadDb();
   // Get the write concern options
   let finalOptions = Object.assign({}, { readPreference: ReadPreference.primary }, options);
   finalOptions = applyWriteConcern(finalOptions, { db }, options);
@@ -640,16 +654,8 @@ function indexInformation(db, name, options, callback) {
  * @param {object} [options] Optional settings. See Db.prototype.listCollections for a list of options.
  * @return {CommandCursor}
  */
-const listCollectionsSchema = {
-  nameOnly: { type: 'boolean', default: false },
-  batchSize: { type: 'number' },
-  readPreference: { type: [ReadPreference, 'object', 'string'] },
-  session: { type: ClientSession },
-  cursorFactory: { overrideOnly: true },
-  optionsValidationLevel: { overrideOnly: true }
-};
 function listCollections(db, filter, options) {
-  const Db = require('../db');
+  let Db = loadDb();
   const find = require('./collection_ops').find;
   options = options || {};
   options = applyDefaults(listCollectionsSchema, options, {});
